Extract numberinput into own file.
