+++ waves/src/components/AssetSelector.tsx
@@ -1,15 +1,8 @@
-import {
-    Center,
-    InputGroup,
-    InputLeftAddon,
-    NumberInput as CUINumberInput,
-    NumberInputField,
-    NumberInputStepper,
-    VStack,
-} from "@chakra-ui/react";
+import { Center, VStack } from "@chakra-ui/react";
 import React, { Dispatch } from "react";
 import { Action, Asset, AssetSide } from "../App";
 import AssetSelect from "./AssetSelect";
+import NumberInput from "./NumberInput";
 
 type StringOrNumber = string | number;
 
@@ -89,62 +82,3 @@ function AssetSelector({ assetSide, type, amount, placement, dispatch }: AssetSe
 }
 
 export default AssetSelector;
-
-interface CustomInputProps {
-    currency: string;
-    value: StringOrNumber;
-    precision: number;
-    step: number;
-    onAmountChange: (val: string) => void;
-    isDisabled: boolean;
-    data_cy: string;
-}
-
-const ASSET_INPUT_LEFT_ADDON_PROPS = {
-    size: "lg",
-    textStyle: "lgGray",
-    w: "15%",
-    h: "3rem",
-    bg: "grey.50",
-    borderRadius: "md",
-    shadow: "md",
-};
-
-const ASSET_INPUT_PROPS = {
-    w: "100%",
-    size: "lg",
-    bg: "#FFFFFF",
-    textStyle: "lgGray",
-    borderRadius: "md",
-    shadow: "md",
-};
-
-const ASSET_INPUT_DISABLED_PROPS = {
-    ...ASSET_INPUT_PROPS,
-    bg: "grey.50",
-};
-
-function NumberInput({ currency, value, onAmountChange, precision, step, isDisabled, data_cy }: CustomInputProps) {
-    const inputProps = isDisabled ? ASSET_INPUT_DISABLED_PROPS : ASSET_INPUT_PROPS;
-    return (
-        <InputGroup>
-            <InputLeftAddon
-                {...ASSET_INPUT_LEFT_ADDON_PROPS}
-                children={currency}
-            />
-            <CUINumberInput
-                {...inputProps}
-                onChange={(valueString, _) => onAmountChange(valueString)}
-                value={value}
-                precision={precision}
-                step={step}
-                isDisabled={isDisabled}
-                min={0}
-                data-cy={`${data_cy}-amount-input`}
-            >
-                <NumberInputField />
-                <NumberInputStepper />
-            </CUINumberInput>
-        </InputGroup>
-    );
-}
